fix #348; tidy old code paths
