+++ src/create/aws/_create-code.js
@@ -23,7 +23,7 @@ let httpLambda = `// @architect/functions enables secure sessions, express-style
 exports.handler = async function http(req) {
   console.log(req)
   return {
-    type: 'text/html; charset=utf8',
+    headers: {'content-type': 'text/html; charset=utf8'},
     body: '<h1>Hello world!</h1>'
   }
 }
