--- src/create/aws/create-http-route/create-route/_errors-html.vtl
@@ -1,4 +1,0 @@
-## Version 1.0.0
-
-#set ($errorMessageObj = $util.parseJson($input.path('$.errorMessage')))
-$errorMessageObj.html
