--- src/create/aws/create-http-route/create-route/_errors-json.vtl
@@ -1,12 +1,0 @@
-## Version 1.0.0
-
-## truncate first half of message
-#set($raw = $input.path('$.errorMessage'))
-#set($len = $input.path('$.errorMessage').length() - 1)
-#set($body = $raw.substring(21, $len))
-## capture everything up to cookie
-#set($start = 0)
-#set($end = $body.indexOf(',"cookie'))
-#set($result = $body.substring($start, $end))
-## print the result
-$result
