--- src/create/aws/create-http-route/create-route/_put-method.js
@@ -1,9 +1,0 @@
-let aws = require('aws-sdk')
-
-module.exports = function putMethod(params, callback) {
-  setTimeout(function _latency() {
-    let gateway = new aws.APIGateway({region: process.env.AWS_REGION})
-    gateway.putMethod(params, callback)
-  }, 1111)
-}
-
