+++ src/create/aws/create-http-route/create-route/_request.vtl
@@ -1,4 +1,4 @@
-## Version 2.0.0
+## Version 2.0.1
 
 #set($inputJson = $util.base64Decode($input.json('$')))
 #if($inputJson.isEmpty())
@@ -13,20 +13,23 @@
   "headers": {
     #foreach($header in $input.params().header.keySet())
     "$header": "$util.escapeJavaScript($input.params().header.get($header))" #if($foreach.hasNext),#end
-
     #end
   },
   "method": "$context.httpMethod",
+  "httpMethod": "$context.httpMethod",
   "params": {
     #foreach($param in $input.params().path.keySet())
     "$param": "$util.escapeJavaScript($input.params().path.get($param))" #if($foreach.hasNext),#end
-
     #end
   },
   "query": {
     #foreach($queryParam in $input.params().querystring.keySet())
     "$queryParam": "$util.escapeJavaScript($input.params().querystring.get($queryParam))" #if($foreach.hasNext),#end
-
+    #end
+  },
+  "queryStringParameters": {
+    #foreach($queryParam in $input.params().querystring.keySet())
+    "$queryParam": "$util.escapeJavaScript($input.params().querystring.get($queryParam))" #if($foreach.hasNext),#end
     #end
   }
 }
